Moved 'stats' to separate module
