+++ application/src/main/java/org/thingsboard/server/ThingsboardInstallApplication.java
@@ -29,7 +29,8 @@ import java.util.Arrays;
 @ComponentScan({"org.thingsboard.server.install",
         "org.thingsboard.server.service.component",
         "org.thingsboard.server.service.install",
-        "org.thingsboard.server.dao"})
+        "org.thingsboard.server.dao",
+        "org.thingsboard.server.common.stats"})
 public class ThingsboardInstallApplication {
 
     private static final String SPRING_CONFIG_NAME_KEY = "--spring.config.name";
