+++ common/pom.xml
@@ -41,6 +41,7 @@
         <module>queue</module>
         <module>transport</module>
         <module>dao-api</module>
+        <module>stats</module>
     </modules>
 
 </project>
