+++ common/queue/src/main/java/org/thingsboard/server/queue/common/DefaultTbQueueResponseTemplate.java
@@ -23,7 +23,7 @@ import org.thingsboard.server.queue.TbQueueHandler;
 import org.thingsboard.server.queue.TbQueueMsg;
 import org.thingsboard.server.queue.TbQueueProducer;
 import org.thingsboard.server.queue.TbQueueResponseTemplate;
-import org.thingsboard.server.common.msg.stats.MessagesStats;
+import org.thingsboard.server.common.stats.MessagesStats;
 
 import java.util.List;
 import java.util.UUID;
