+++ common/stats/src/main/java/org/thingsboard/server/common/stats/DefaultMessagesStats.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.thingsboard.server.common.msg.stats;
+package org.thingsboard.server.common.stats;
 
 public class DefaultMessagesStats implements MessagesStats {
     private final StatsCounter totalCounter;
