+++ common/stats/src/main/java/org/thingsboard/server/common/stats/DefaultStatsFactory.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.thingsboard.server.common.msg.stats;
+package org.thingsboard.server.common.stats;
 
 import io.micrometer.core.instrument.Counter;
 import io.micrometer.core.instrument.MeterRegistry;
